chore: tighen up new testing configuration
