+++ tests/data_preparation/sample_category_data.js
@@ -1,0 +1,7 @@
+const sample_category_data = function () {
+  return {
+    item1: {},
+  };
+};
+
+module.exports = sample_category_data;
